Wire up upgrade orchestrator in engine
