+++ pkg/protocol/engine/upgrade/signalingupgradeorchestrator/snapshot.go
@@ -1,4 +1,4 @@
-package upgrade
+package signalingupgradeorchestrator
 
 import (
 	"io"
